pyproject.toml
[project]
name = "yaml-update"
version = "1.0.0"
description = "GitHub Action to automate updating YAML files"
readme = "readme.md"
license = "MIT"
requires-python = ">=3.13"
dependencies = [
    "ruamel.yaml>=0.18",
    "PyGithub>=2.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.8",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py313"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.coverage.run]
source = ["yaml_update"]

[tool.coverage.report]
show_missing = true
fail_under = 80
